requirements.txt
langchain==0.2.1
langchain_community==0.2.1
langchain_openai==0.1.7
nest_asyncio==1.6.0
openai==1.30.4
pinecone==4.0.0
PyDrive2==1.19.0
python-dotenv==1.0.1
python_docx==1.1.2
streamlit==1.31.1
streamlit_chat==0.1.1
streamlit_option_menu==0.3.12
urllib3==1.26.18
